+++ sdk-libs/photon-api/src/models/_get_compressed_token_accounts_by_owner_post_request_params.rs
@@ -16,9 +16,9 @@ pub struct GetCompressedTokenAccountsByOwnerPostRequestParams {
     #[serde(rename = "cursor", default, skip_serializing_if = "Option::is_none")]
     pub cursor: Option<String>,
     #[serde(rename = "limit", default, skip_serializing_if = "Option::is_none")]
-    pub limit: Option<u16>,
+    pub limit: Option<i32>,
     /// A Solana public key represented as a base58 string.
-    #[serde(rename = "mint", default, skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "mint", default,  skip_serializing_if = "Option::is_none")]
     pub mint: Option<String>,
     /// A Solana public key represented as a base58 string.
     #[serde(rename = "owner")]
